add LoaderScheduler
