+++ linen-modern/src/main/scala/x7c1/linen/modern/init/unread/UnreadItemsDelegatee.scala
@@ -64,7 +64,7 @@ class UnreadItemsDelegatee(
         false
     }
   }
-  private def setupLoaderSchedule(): Unit ={
+  private def setupLoaderSchedule(): Unit = {
     clientAccount foreach { account =>
       ServiceCaller.using[UpdaterMethods].
         startService(activity, activity getClassOf Updater){
