+++ linen-scene/src/main/scala/x7c1/linen/scene/updater/UpdaterMethods.scala
@@ -1,18 +1,17 @@
 package x7c1.linen.scene.updater
 
-import java.util.Calendar
-
-import android.app.{PendingIntent, AlarmManager, Service}
-import android.content.{Context, Intent}
+import android.app.Service
+import android.content.Intent
 import x7c1.linen.database.control.DatabaseHelper
-import x7c1.linen.glue.service.{ServiceLabel, ServiceControl}
+import x7c1.linen.glue.service.ServiceControl
 import x7c1.linen.repository.date.Date
 import x7c1.linen.repository.dummy.{DummyFactory, TraceableQueue}
 import x7c1.linen.repository.loader.crawling.{Implicits, SourceInspector}
 import x7c1.linen.repository.loader.schedule.LoaderSchedule
 import x7c1.linen.repository.preset.PresetFactory
 import x7c1.linen.repository.source.setting.{SettingSource, SettingSourceAccessorFactory}
-import x7c1.wheat.macros.intent.{IntentFactory, ExtraNotFound, IntentExpander}
+import x7c1.linen.scene.loader.crawling.LoaderScheduler
+import x7c1.wheat.macros.intent.{ExtraNotFound, IntentExpander}
 import x7c1.wheat.macros.logger.Log
 import x7c1.wheat.modern.formatter.ThrowableFormatter.format
 import x7c1.wheat.modern.patch.TaskAsync.async
@@ -48,32 +47,13 @@ class UpdaterMethods(
   def createDummySources(channelIds: Seq[Long]) = async {
     Log info s"$channelIds"
   }
-  def setupLoaderSchedule(accountId: Long): Unit = {
+  def setupLoaderSchedule(accountId: Long): Unit = Future {
     Log error s"[init]"
 
-    helper.selectorOf[LoaderSchedule].findPresetSchedule(accountId) map { schedule =>
-      schedule.startRanges.toSeq.foreach { range =>
-        Log info "h:" + range.from.hour.value
-        Log info "m:" + range.from.minute.value
-
-        val manager = service.getSystemService(Context.ALARM_SERVICE).asInstanceOf[AlarmManager]
-
-        val pendingIntent = {
-          val intent = IntentFactory.using[UpdaterMethods].
-            create(service, service getClassOf ServiceLabel.Updater){
-              _.loadFromSchedule(schedule.scheduleId)
-            }
-
-          PendingIntent.getService(service, 0, intent, 0)
-        }
-        val msec = {
-          val calendar = Calendar.getInstance()
-          calendar setTimeInMillis System.currentTimeMillis()
-          calendar.add(Calendar.SECOND, 10)
-          calendar.getTimeInMillis
-        }
-        manager.set(AlarmManager.RTC_WAKEUP, msec, pendingIntent)
-      }
+    helper.selectorOf[LoaderSchedule] findPresetSchedule accountId via {
+      case Right(Some(schedule)) => LoaderScheduler(service) createOrUpdate schedule
+      case Right(None) => Log error s"preset schedule not found"
+      case Left(e) => Log error format(e){"[failed]"}
     }
   }
   def loadFromSchedule(scheduleId: Long): Unit = {
